+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 This is used for recording Invader's changes. This changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Untagged]
+### Fixed
+- invader: Physics tags' center of mass and mass point density/mass are now calculated on
+  build and are cache-only.
+
 ## [0.53.7] - 2024-06-16
 ### Fixed
 - invader-build: Clamp particle animation rate the same way as tool.exe
@@ -2546,4 +2551,4 @@ image as one bitmap
 [0.53.6]: https://github.com/SnowyMouse/invader/compare/0.53.5...0.53.6
 [0.53.7]: https://github.com/SnowyMouse/invader/compare/0.53.6...0.53.7
 
-[Untagged]: https://github.com/SnowyMouse/invader/compare/0.53.6...master
+[Untagged]: https://github.com/SnowyMouse/invader/compare/0.53.7...master